Finalizado upload via s3

+++ src/controllers/PostController.js
@@ -11,28 +11,45 @@ module.exports = {
 
   async store(req, res) {
     const { author, place, description, hashtags } = req.body;
-    const { size, location: url = "", filename: image } = req.file;
+    const { size, key, location: url = "", filename: image } = req.file;
+    let postObj = {};
 
-    const [name] = image.split(".");
-    let fileName = `${name}.jpg`;
-    fileName = fileName.replace(" ", "");
+    if (typeof key === "undefined") {
+      await sharp(req.file.path)
+        .resize(500)
+        .jpeg({ quality: 70 })
+        .toFile(path.resolve(req.file.destination, "resized", fileName));
 
-    await sharp(req.file.path)
-      .resize(500)
-      .jpeg({ quality: 70 })
-      .toFile(path.resolve(req.file.destination, "resized", fileName));
+      fs.unlinkSync(req.file.path); // Apaga do diretório uploads (fotos sem resize)
 
-    fs.unlinkSync(req.file.path); // Apaga do diretório uploads (fotos sem resize)
+      const [name] = image.split(".");
+      let fileName = `${name}.jpg`;
+      fileName = fileName.replace(" ", "");
 
-    const post = await Post.create({
-      author,
-      place,
-      description,
-      hashtags,
-      image: fileName,
-      size,
-      url
-    });
+      postObj = {
+        author,
+        place,
+        description,
+        hashtags,
+        image: fileName,
+        key: fileName,
+        size,
+        url
+      };
+    } else {
+      postObj = {
+        author,
+        place,
+        description,
+        hashtags,
+        image: key,
+        key,
+        size,
+        url
+      };
+    }
+
+    const post = await Post.create(postObj);
 
     req.io.emit("post", post);
